refactor(wire): split `node.rs` by behavior
So far `node.rs` has been a catch-all file, since it was containing the definition for types, macros, and methods for anything needed during the `UtreexoNode` operation. It was already on the 2,250 lines, which makes it hard to reason about and test.

I think there are roughly four kind of behaviors:
- Peer management (including inflight requests, addnode, message times, etc.)
- Connectivity (opening connections, bootstrapping via DNS/seeds, and maintenance)
- Block pipeline (requesting blocks/proofs and processing them)
- Handling user requests

So I created four files under the new `node/` folder for these behaviors, along with the three node contexts.

The only code changes were:
- Making `broadcast_to_peers`, `resolve_connect_host`, `get_port` and `to_addr_v2` `pub(crate)`
- Moving the `Default` impl for `RunningNode` to its own file
- Renaming `maybe_use_hadcoded_addresses` to `maybe_use_hardcoded_addresses` (typo).

Hence, this change is move-only in terms of the method bodies (should be easy to verify).
