+++ crates/floresta-node/src/florestad.rs
@@ -28,8 +28,8 @@ use floresta_watch_only::kv_database::KvDatabase;
 use floresta_watch_only::AddressCache;
 use floresta_wire::address_man::AddressMan;
 use floresta_wire::mempool::Mempool;
+use floresta_wire::node::running_ctx::RunningNode;
 use floresta_wire::node::UtreexoNode;
-use floresta_wire::running_node::RunningNode;
 use floresta_wire::UtreexoNodeConfig;
 use rcgen::BasicConstraints;
 use rcgen::CertificateParams;
