+++ crates/floresta-wire/src/p2p_wire/mod.rs
@@ -95,16 +95,13 @@ impl Default for UtreexoNodeConfig {
 
 pub mod address_man;
 pub mod block_proof;
-pub mod chain_selector;
 pub mod error;
 pub mod mempool;
 pub mod node;
 pub mod node_context;
 pub mod node_interface;
 pub mod peer;
-pub mod running_node;
 pub mod socks;
-pub mod sync_node;
 #[cfg(test)]
 #[doc(hidden)]
 pub mod tests;
